shared/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
httpx==0.27.0

# PostgreSQL async
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Alembic for migrations
alembic==1.13.1

# JWT and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pyotp==2.9.0
qrcode[pil]==8.0

# Pydantic for validation
pydantic==2.5.3
pydantic[email]==2.5.3
email-validator==2.1.0

# Rate limiting
slowapi==0.1.9
limits==3.10.1
redis==5.0.1

# Telegram Bot (aiogram 3.x)
aiogram==3.4.1
aiohttp==3.9.3

# Payment Gateway SDKs
# Note: heleket SDK is optional, can use httpx directly for Helket API
# heleket>=1.0.0

# ClickHouse client for analytical queries
clickhouse-connect>=0.7.0
